refactor: transalte locker actions
