--- src/actions/lockers/location/read.ts
@@ -1,4 +1,0 @@
-import { action } from '@/actions/action'
-import { LockerLocationMethods } from '@/services/lockers/locations/methods'
-
-export const readAllLockerLocationsAction = action(LockerLocationMethods.readAll)
