+++ src/actions/lockers/reservations.ts
@@ -3,4 +3,5 @@
 import { action } from '@/actions/action'
 import { LockerReservationMethods } from '@/services/lockers/reservations/methods'
 
+export const updateLockerReservationAction = action(LockerReservationMethods.update)
 export const createLockerReservationAction = action(LockerReservationMethods.create)
