--- src/actions/lockers/reservations/update.ts
@@ -1,6 +1,0 @@
-'use server'
-
-import { action } from '@/actions/action'
-import { LockerReservationMethods } from '@/services/lockers/reservations/methods'
-
-export const updateLockerReservationAction = action(LockerReservationMethods.update)
