+++ src/app/lockers/[id]/UpdateLockerReservationForm.tsx
@@ -1,12 +1,12 @@
 'use client'
-import { updateLockerReservationAction } from '@/actions/lockers/reservations/update'
+import { updateLockerReservationAction } from '@/actions/lockers/reservations'
 import Form from '@/components/Form/Form'
 import { SelectString } from '@/components/UI/Select'
 import DateInput from '@/components/UI/DateInput'
 import Checkbox from '@/components/UI/Checkbox'
+import { bindParams } from '@/actions/bind'
 import { useState } from 'react'
 import { useRouter } from 'next/navigation'
-import { bindParams } from '@/actions/bind'
 
 type PropTypes = {
     reservationId: number,
