+++ src/app/lockers/[id]/page.tsx
@@ -3,7 +3,7 @@ import LockerNotFound from './LockerNotFound'
 import CreateLockerReservationForm from './CreateLockerReservationForm'
 import UpdateLockerReservationForm from './UpdateLockerReservationForm'
 import PageWrapper from '@/components/PageWrapper/PageWrapper'
-import { readLockerAction } from '@/actions/lockers/read'
+import { readLockerAction } from '@/actions/lockers/lockers'
 import { getUser } from '@/auth/getUser'
 import { checkGroupValidity, inferGroupName, readGroupsOfUser } from '@/services/groups/read'
 
@@ -18,7 +18,7 @@ export default async function Locker({ params }: PropTypes) {
     const { user, status, authorized } = await getUser({
         userRequired: true,
         shouldRedirect: true,
-        requiredPermissions: [['LOCKER_READ']],
+        requiredPermissions: [['LOCKER_USE']],
     })
     if (!authorized) return Error(status)
