+++ src/services/lockers/authers.ts
@@ -1,7 +1,7 @@
 import { RequirePermission } from '@/auth/auther/RequirePermission'
 
 export namespace LockerAuthers {
-    export const create = RequirePermission.staticFields({ permission: 'LOCKER_CREATE' })
-    export const read = RequirePermission.staticFields({ permission: 'LOCKER_READ' })
-    export const readPage = RequirePermission.staticFields({ permission: 'LOCKER_READ' })
+    export const create = RequirePermission.staticFields({ permission: 'LOCKER_ADMIN' })
+    export const read = RequirePermission.staticFields({ permission: 'LOCKER_USE' })
+    export const readPage = RequirePermission.staticFields({ permission: 'LOCKER_USE' })
 }
