+++ src/services/prismaCall.ts
@@ -1,7 +1,7 @@
 import { ServerError, Smorekopp } from './error'
+import logger from '@/lib/logger'
 import { Prisma } from '@prisma/client'
 import type { ServerErrorCode } from './error'
-import logger from '@/lib/logger'
 
 const errorMessagesMap: { [key: string]: [ServerErrorCode, string] } = {
     P2002: ['DUPLICATE', 'duplicate entry'],
@@ -26,13 +26,13 @@ export async function prismaCall<T>(call: () => T | Promise<T>): Promise<T> {
         }
 
         if (!(error instanceof Prisma.PrismaClientKnownRequestError)) {
-            logger.error("Unknown error:", error)
+            logger.error('Unknown error:', error)
             throw new ServerError('UNKNOWN ERROR', 'unknown error')
         }
 
         const pError = errorMessagesMap[error.code]
         if (pError) throw new ServerError(pError[0], pError[1])
-        logger.error("Unknown prisma error:", error)
+        logger.error('Unknown prisma error:', error)
         throw new ServerError('UNKNOWN ERROR', 'unknown prisma error')
     }
 }
